refactor: WinningLottoTest 분리
