+++ src/test/java/domain/lotto/LottoNumbersTest.java
@@ -79,72 +79,4 @@ public class LottoNumbersTest {
         assertThat(lottoNumbers).isEqualTo(sortedLottoNumbers);
     }
 
-    @ParameterizedTest
-    @MethodSource("provideLottoRank")
-    void 로또_결과에_맞는_랭크를_반환한다(WinningLotto winningLotto, Rank expected) {
-        // given
-        LottoNumbers lottoNumbers = new LottoNumbers(lottoNumberList);
-
-        // when, then
-        assertThat(winningLotto.getRank(lottoNumbers)).isEqualTo(Optional.ofNullable(expected));
-    }
-
-    private static Stream<Arguments> provideLottoRank() {
-        LottoNumbers firstPlaceLottoNumbers = new LottoNumbers(List.of(
-                LottoNumber.of(31),
-                LottoNumber.of(18),
-                LottoNumber.of(5),
-                LottoNumber.of(22),
-                LottoNumber.of(41),
-                LottoNumber.of(9)
-        ));
-
-        LottoNumbers secondOrThirdPlaceLottoNumbers = new LottoNumbers(List.of(
-                LottoNumber.of(31),
-                LottoNumber.of(18),
-                LottoNumber.of(5),
-                LottoNumber.of(22),
-                LottoNumber.of(41),
-                LottoNumber.of(1)
-        ));
-
-        LottoNumbers fourthPlaceLottoNumbers = new LottoNumbers(List.of(
-                LottoNumber.of(31),
-                LottoNumber.of(18),
-                LottoNumber.of(5),
-                LottoNumber.of(22),
-                LottoNumber.of(2),
-                LottoNumber.of(1)
-        ));
-
-        LottoNumbers fifthPlaceLottoNumbers = new LottoNumbers(List.of(
-                LottoNumber.of(31),
-                LottoNumber.of(18),
-                LottoNumber.of(5),
-                LottoNumber.of(3),
-                LottoNumber.of(2),
-                LottoNumber.of(1)
-        ));
-
-        LottoNumbers noPlaceLottoNumbers = new LottoNumbers(List.of(
-                LottoNumber.of(5),
-                LottoNumber.of(6),
-                LottoNumber.of(4),
-                LottoNumber.of(3),
-                LottoNumber.of(2),
-                LottoNumber.of(1)
-        ));
-
-        LottoNumber bonusBallForSecondPlace = LottoNumber.of(9);
-        LottoNumber bonusBallNotAffect = LottoNumber.of(45);
-
-        return Stream.of(
-                Arguments.arguments(new WinningLotto(firstPlaceLottoNumbers, bonusBallNotAffect), Rank.FIRST_PLACE),
-                Arguments.arguments(new WinningLotto(secondOrThirdPlaceLottoNumbers, bonusBallForSecondPlace), Rank.SECOND_PLACE),
-                Arguments.arguments(new WinningLotto(secondOrThirdPlaceLottoNumbers, bonusBallNotAffect), Rank.THIRD_PLACE),
-                Arguments.arguments(new WinningLotto(fourthPlaceLottoNumbers, bonusBallNotAffect), Rank.FOURTH_PLACE),
-                Arguments.arguments(new WinningLotto(fifthPlaceLottoNumbers, bonusBallNotAffect), Rank.FIFTH_PLACE),
-                Arguments.arguments(new WinningLotto(noPlaceLottoNumbers, bonusBallNotAffect), null)
-        );
-    }
 }
