+++ src/test/java/domain/lotto/WinningLottoTest.java
@@ -1,45 +1,43 @@
 package domain.lotto;
 
 import domain.rank.Rank;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
+import java.util.Optional;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class LottoTest {
+public class WinningLottoTest {
+    private List<LottoNumber> lottoNumberList;
 
-    @ParameterizedTest
-    @MethodSource("provideLotto")
-    void 여러_장의_로또_당첨_등수를_계산한다(List<LottoNumbers> lottoNumbersList, LottoNumber bonusBall, List<Integer> expectedRankCountList) {
-        // given
-        Lotto lotto = new Lotto(lottoNumbersList);
-        LottoNumbers winLottoNumbers = new LottoNumbers(List.of(
+    @BeforeEach
+    void setUp() {
+        lottoNumberList = Arrays.asList(
                 LottoNumber.of(31),
                 LottoNumber.of(18),
                 LottoNumber.of(5),
                 LottoNumber.of(22),
                 LottoNumber.of(41),
                 LottoNumber.of(9)
-        ));
+        );
+    }
+    @ParameterizedTest
+    @MethodSource("provideLottoRank")
+    void 로또_결과에_맞는_랭크를_반환한다(WinningLotto winningLotto, Rank expected) {
+        // given
+        LottoNumbers lottoNumbers = new LottoNumbers(lottoNumberList);
 
-        // when
-        Map<Rank, Integer> rankMap = lotto.rankEachLotto(new WinningLotto(winLottoNumbers, bonusBall));
-
-        // then
-        List<Integer> actualRankCountList = Arrays.stream(Rank.values())
-                .map(rankMap::get)
-                .collect(Collectors.toList());
-        assertThat(actualRankCountList).isEqualTo(expectedRankCountList);
+        // when, then
+        assertThat(winningLotto.getRank(lottoNumbers)).isEqualTo(expected);
     }
 
-    private static Stream<Arguments> provideLotto() {
+    private static Stream<Arguments> provideLottoRank() {
         LottoNumbers firstPlaceLottoNumbers = new LottoNumbers(List.of(
                 LottoNumber.of(31),
                 LottoNumber.of(18),
@@ -49,7 +47,7 @@ public class LottoTest {
                 LottoNumber.of(9)
         ));
 
-        LottoNumbers secondorThirdPlaceLottoNumbers = new LottoNumbers(List.of(
+        LottoNumbers secondOrThirdPlaceLottoNumbers = new LottoNumbers(List.of(
                 LottoNumber.of(31),
                 LottoNumber.of(18),
                 LottoNumber.of(5),
@@ -85,11 +83,16 @@ public class LottoTest {
                 LottoNumber.of(1)
         ));
 
-        List<LottoNumbers> lottoNumbersList = List.of(firstPlaceLottoNumbers, secondorThirdPlaceLottoNumbers, fourthPlaceLottoNumbers, fifthPlaceLottoNumbers, noPlaceLottoNumbers);
+        LottoNumber bonusBallForSecondPlace = LottoNumber.of(9);
+        LottoNumber bonusBallNotAffect = LottoNumber.of(45);
 
         return Stream.of(
-                Arguments.arguments(lottoNumbersList, LottoNumber.of(1), List.of(1, 1, 0, 1, 1, 1)),
-                Arguments.arguments(lottoNumbersList, LottoNumber.of(33), List.of(1, 0, 1, 1, 1, 1))
+                Arguments.arguments(new WinningLotto(firstPlaceLottoNumbers, bonusBallNotAffect), Rank.FIRST_PLACE),
+                Arguments.arguments(new WinningLotto(secondOrThirdPlaceLottoNumbers, bonusBallForSecondPlace), Rank.SECOND_PLACE),
+                Arguments.arguments(new WinningLotto(secondOrThirdPlaceLottoNumbers, bonusBallNotAffect), Rank.THIRD_PLACE),
+                Arguments.arguments(new WinningLotto(fourthPlaceLottoNumbers, bonusBallNotAffect), Rank.FOURTH_PLACE),
+                Arguments.arguments(new WinningLotto(fifthPlaceLottoNumbers, bonusBallNotAffect), Rank.FIFTH_PLACE),
+                Arguments.arguments(new WinningLotto(noPlaceLottoNumbers, bonusBallNotAffect), Rank.NOTHING)
         );
     }
 }
